fix bug in non assignments
